development-web-app/development-backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
pytz==2024.2
sqlparse==0.5.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
inject~=5.2.1
factory_boy==3.3.1
pytest-django==4.9.0
requests==2.32.3
scikit-learn~=1.5.2
xgboost~=2.1.2
joblib==1.4.2
pandas==2.2.3
ortools==9.11.4210
django-secured-fields==0.4.4
djangorestframework-dataclasses==1.3.1